+++ product_code/examreg_service/api/src/main/scala/de/upb/cs/uc4/examreg/model/JsonExamRegNameList.scala
@@ -1,9 +1,0 @@
-package de.upb.cs.uc4.examreg.model
-
-import play.api.libs.json.{ Format, Json }
-
-case class JsonExamRegNameList(examinationRegulations: Seq[String])
-
-object JsonExamRegNameList {
-  implicit val format: Format[JsonExamRegNameList] = Json.format
-}