+++ product_code/examreg_service/api/src/main/scala/de/upb/cs/uc4/examreg/model/JsonExaminationRegulations.scala
@@ -1,9 +1,0 @@
-package de.upb.cs.uc4.examreg.model
-
-import play.api.libs.json.{ Format, Json }
-
-case class JsonExaminationRegulations(examinationRegulations: Seq[ExaminationRegulation])
-
-object JsonExaminationRegulations {
-  implicit val format: Format[JsonExaminationRegulations] = Json.format
-}