+++ product_code/examreg_service/impl/src/main/scala/de/upb/cs/uc4/examreg/impl/ExamregServiceImpl.scala
@@ -10,11 +10,11 @@ import de.upb.cs.uc4.examreg.api.ExamregService
 import de.upb.cs.uc4.examreg.impl.actor.ExamregState
 import de.upb.cs.uc4.examreg.impl.commands.{ ExamregCommand, GetExamreg }
 import de.upb.cs.uc4.examreg.impl.readside.{ ExamregDatabase, ExamregEventProcessor }
-import de.upb.cs.uc4.examreg.model.{ ExaminationRegulation, JsonExamRegNameList, JsonExaminationRegulations }
+import de.upb.cs.uc4.examreg.model.{ ExaminationRegulation, Module }
 import de.upb.cs.uc4.shared.server.ServiceCallFactory._
 
-import scala.concurrent.{ ExecutionContext, Future }
 import scala.concurrent.duration._
+import scala.concurrent.{ ExecutionContext, Future }
 
 /** Implementation of the ExamregService */
 class ExamregServiceImpl(clusterSharding: ClusterSharding, readSide: ReadSide,
@@ -27,8 +27,11 @@ class ExamregServiceImpl(clusterSharding: ClusterSharding, readSide: ReadSide,
 
   implicit val timeout: Timeout = Timeout(15.seconds)
 
+  // TODO All ExamReg Tests
+  // TODO Default ExamRegs
+
   /** @inheritdoc */
-  override def getExaminationRegulations(regulations: Option[String]): ServiceCall[NotUsed, JsonExaminationRegulations] =
+  override def getExaminationRegulations(regulations: Option[String], active: Option[Boolean]): ServiceCall[NotUsed, Seq[ExaminationRegulation]] =
     ServiceCall { _ =>
       database.getAll
         .map(names => names
@@ -36,17 +39,36 @@ class ExamregServiceImpl(clusterSharding: ClusterSharding, readSide: ReadSide,
         .flatMap(seq => Future.sequence(seq)
           .map(seq => seq
             .filter(opt => opt.isDefined) //Filter every not existing examination regulation
-            .map(opt => opt.get)))
-        .map(examregs => JsonExaminationRegulations(examregs))
+            .map(opt => opt.get)
+            .filter(examReg => regulations.isEmpty || regulations.get.split(",").contains(examReg.name))
+            .filter(examReg => active.isEmpty || active.get == examReg.active)))
+
     }
 
   /** @inheritdoc */
-  override def getExaminationRegulationsNames: ServiceCall[NotUsed, JsonExamRegNameList] = ServiceCall { _ =>
-    database.getAll.map(seq => JsonExamRegNameList(seq))
+  override def getModules(moduleIds: Option[String], active: Option[Boolean]): ServiceCall[NotUsed, Seq[Module]] = ServiceCall {
+    _ =>
+      getExaminationRegulations(None, None).invoke().map {
+        _.filter(active.isEmpty || _.active == active.get)
+          .flatMap(x => x.modules).distinct
+          .filter(module => moduleIds.isEmpty || moduleIds.contains(module.id))
+      }
+  }
+
+  /** @inheritdoc */
+  override def getExaminationRegulationsNames(active: Option[Boolean]): ServiceCall[NotUsed, Seq[String]] = ServiceCall { _ =>
+    active match {
+      case None => database.getAll
+      case Some(isActive) => getExaminationRegulations(None, None).invoke().map {
+        _.filter(_.active == isActive).map(_.name)
+      }
+    }
+
   }
 
   /** @inheritdoc */
   override def allowedMethodsGET: ServiceCall[NotUsed, Done] = allowedMethodsCustom("GET")
 
   override def allowVersionNumber: ServiceCall[NotUsed, Done] = allowedMethodsCustom("GET")
+
 }